requirements.txt
requests==2.31.0
python-dotenv==1.0.0
groq==0.37.0
streamlit==1.29.0
